// packages/protocol/typescript/generated/agent_start/runtime/v1/worktree_pb.d.ts
export declare const WorktreeServiceArchiveRequestSchema: GenMessage<WorktreeServiceArchiveRequest>;
// ...
export declare const WorktreeServiceArchiveResponseSchema: GenMessage<WorktreeServiceArchiveResponse>;
// ...
export declare const WorktreeServiceListRequestSchema: GenMessage<WorktreeServiceListRequest>;
// ...
export declare const WorktreeServiceListResponseSchema: GenMessage<WorktreeServiceListResponse>;
// ...
export declare const WorktreeServiceCreateRequestSchema: GenMessage<WorktreeServiceCreateRequest>;
// ...
export declare const WorktreeServiceCreateResponseSchema: GenMessage<WorktreeServiceCreateResponse>;
// ...
export declare const WorktreeServiceListArchivesRequestSchema: GenMessage<WorktreeServiceListArchivesRequest>;
// ...
export declare const WorktreeServiceListArchivesResponseSchema: GenMessage<WorktreeServiceListArchivesResponse>;
// ...
export declare const WorktreeServiceRestoreRequestSchema: GenMessage<WorktreeServiceRestoreRequest>;
// ...
export declare const WorktreeServiceRestoreResponseSchema: GenMessage<WorktreeServiceRestoreResponse>;
// ...
export declare const WorktreeServicePsRequestSchema: GenMessage<WorktreeServicePsRequest>;
// ...
export declare const WorktreeServicePsResponseSchema: GenMessage<WorktreeServicePsResponse>;
// ...
export declare const WorktreeServiceShowRequestSchema: GenMessage<WorktreeServiceShowRequest>;
// ...
export declare const WorktreeServiceShowResponseSchema: GenMessage<WorktreeServiceShowResponse>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServiceSleepRequest
 */
export declare type WorktreeServiceSleepRequest = Message<"agent_start.runtime.v1.WorktreeServiceSleepRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServiceSleepRequest.
 * Use `create(WorktreeServiceSleepRequestSchema)` to create a new message.
 */
export declare const WorktreeServiceSleepRequestSchema: GenMessage<WorktreeServiceSleepRequest>;
// ...
export declare const WorktreeServiceSleepResponseSchema: GenMessage<WorktreeServiceSleepResponse>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServiceActivateRequest
 */
export declare type WorktreeServiceActivateRequest = Message<"agent_start.runtime.v1.WorktreeServiceActivateRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: bool notify_clients = 2;
   */
  notifyClients: boolean;
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServiceActivateRequest.
 * Use `create(WorktreeServiceActivateRequestSchema)` to create a new message.
 */
export declare const WorktreeServiceActivateRequestSchema: GenMessage<WorktreeServiceActivateRequest>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServiceActivateResponse
 */
export declare type WorktreeServiceActivateResponse = Message<"agent_start.runtime.v1.WorktreeServiceActivateResponse"> & {
  /**
   * @generated from field: string repo_id = 1;
   */
  repoId: string;

  /**
   * @generated from field: string worktree_id = 2;
   */
  worktreeId: string;

  /**
   * @generated from field: bool activated = 3;
   */
  activated: boolean;

  /**
   * @generated from field: agent_start.runtime.v1.WorktreeSleepingAgentWake sleeping_agent_wake = 4;
   */
  sleepingAgentWake: WorktreeSleepingAgentWake;
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServiceActivateResponse.
 * Use `create(WorktreeServiceActivateResponseSchema)` to create a new message.
 */
export declare const WorktreeServiceActivateResponseSchema: GenMessage<WorktreeServiceActivateResponse>;
// ...
export declare type WorktreeServiceRemoveRequest = Message<"agent_start.runtime.v1.WorktreeServiceRemoveRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: int64 expected_revision = 2;
   */
  expectedRevision: bigint;

  /**
   * @generated from field: bool force = 3;
   */
  force: boolean;

  /**
   * @generated from field: bool run_hooks = 4;
   */
  runHooks: boolean;
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServiceRemoveRequest.
 * Use `create(WorktreeServiceRemoveRequestSchema)` to create a new message.
 */
export declare const WorktreeServiceRemoveRequestSchema: GenMessage<WorktreeServiceRemoveRequest>;
// ...
export declare const WorktreeServiceRemoveResponseSchema: GenMessage<WorktreeServiceRemoveResponse>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServiceForceDeleteBranchRequest
 */
export declare type WorktreeServiceForceDeleteBranchRequest = Message<"agent_start.runtime.v1.WorktreeServiceForceDeleteBranchRequest"> & {
  /**
   * @generated from field: string worktree = 1;
   */
  worktree: string;

  /**
   * @generated from field: string branch_name = 2;
   */
  branchName: string;

  /**
   * @generated from field: string expected_head = 3;
   */
  expectedHead: string;
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServiceForceDeleteBranchRequest.
 * Use `create(WorktreeServiceForceDeleteBranchRequestSchema)` to create a new message.
 */
export declare const WorktreeServiceForceDeleteBranchRequestSchema: GenMessage<WorktreeServiceForceDeleteBranchRequest>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServiceForceDeleteBranchResponse
 */
export declare type WorktreeServiceForceDeleteBranchResponse = Message<"agent_start.runtime.v1.WorktreeServiceForceDeleteBranchResponse"> & {
  /**
   * @generated from field: bool deleted = 1;
   */
  deleted: boolean;
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServiceForceDeleteBranchResponse.
 * Use `create(WorktreeServiceForceDeleteBranchResponseSchema)` to create a new message.
 */
export declare const WorktreeServiceForceDeleteBranchResponseSchema: GenMessage<WorktreeServiceForceDeleteBranchResponse>;
// ...
export declare const WorktreeServiceSetRequestSchema: GenMessage<WorktreeServiceSetRequest>;
// ...
export declare const WorktreeServiceSetResponseSchema: GenMessage<WorktreeServiceSetResponse>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServicePersistSortOrderRequest
 */
export declare type WorktreeServicePersistSortOrderRequest = Message<"agent_start.runtime.v1.WorktreeServicePersistSortOrderRequest"> & {
  /**
   * @generated from field: repeated string ordered_ids = 1;
   */
  orderedIds: string[];
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServicePersistSortOrderRequest.
 * Use `create(WorktreeServicePersistSortOrderRequestSchema)` to create a new message.
 */
export declare const WorktreeServicePersistSortOrderRequestSchema: GenMessage<WorktreeServicePersistSortOrderRequest>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServicePersistSortOrderResponse
 */
export declare type WorktreeServicePersistSortOrderResponse = Message<"agent_start.runtime.v1.WorktreeServicePersistSortOrderResponse"> & {
  /**
   * @generated from field: uint32 updated = 1;
   */
  updated: number;
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServicePersistSortOrderResponse.
 * Use `create(WorktreeServicePersistSortOrderResponseSchema)` to create a new message.
 */
export declare const WorktreeServicePersistSortOrderResponseSchema: GenMessage<WorktreeServicePersistSortOrderResponse>;
// ...
export declare const WorktreeServiceDetectedListRequestSchema: GenMessage<WorktreeServiceDetectedListRequest>;
// ...
export declare const WorktreeServiceDetectedListResponseSchema: GenMessage<WorktreeServiceDetectedListResponse>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServiceLineageListRequest
 */
export declare type WorktreeServiceLineageListRequest = Message<"agent_start.runtime.v1.WorktreeServiceLineageListRequest"> & {
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServiceLineageListRequest.
 * Use `create(WorktreeServiceLineageListRequestSchema)` to create a new message.
 */
export declare const WorktreeServiceLineageListRequestSchema: GenMessage<WorktreeServiceLineageListRequest>;
// ...
export declare const WorktreeServiceLineageListResponseSchema: GenMessage<WorktreeServiceLineageListResponse>;

/**
 * @generated from message agent_start.runtime.v1.WorktreeServicePrefetchCreateBaseRequest
 */
export declare type WorktreeServicePrefetchCreateBaseRequest = Message<"agent_start.runtime.v1.WorktreeServicePrefetchCreateBaseRequest"> & {
  /**
   * @generated from field: string repo = 1;
   */
  repo: string;

  /**
   * @generated from field: optional string base_branch = 2;
   */
  baseBranch?: string | undefined;
};

/**
 * Describes the message agent_start.runtime.v1.WorktreeServicePrefetchCreateBaseRequest.
 * Use `create(WorktreeServicePrefetchCreateBaseRequestSchema)` to create a new message.
 */
export declare const WorktreeServicePrefetchCreateBaseRequestSchema: GenMessage<WorktreeServicePrefetchCreateBaseRequest>;
// ...
export declare const WorktreeServicePrefetchCreateBaseResponseSchema: GenMessage<WorktreeServicePrefetchCreateBaseResponse>;
// ...
export declare const WorktreeServiceResolvePrBaseRequestSchema: GenMessage<WorktreeServiceResolvePrBaseRequest>;
// ...
export declare const WorktreeServiceResolvePrBaseResponseSchema: GenMessage<WorktreeServiceResolvePrBaseResponse>;
// ...
export declare const WorktreeServiceBranchRenameFailureOutputRequestSchema: GenMessage<WorktreeServiceBranchRenameFailureOutputRequest>;
// ...
export declare const WorktreeServiceBranchRenameFailureOutputResponseSchema: GenMessage<WorktreeServiceBranchRenameFailureOutputResponse>;
// ...
export declare const WorktreeServiceSubscribeStateEventsRequestSchema: GenMessage<WorktreeServiceSubscribeStateEventsRequest>;
// ...
export declare const WorktreeServiceSubscribeStateEventsResponseSchema: GenMessage<WorktreeServiceSubscribeStateEventsResponse>;
// ...
export declare const WorktreeService: GenService<{
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Archive
   */
  archive: {
    methodKind: "unary";
    input: typeof WorktreeServiceArchiveRequestSchema;
    output: typeof WorktreeServiceArchiveResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.List
   */
  list: {
    methodKind: "unary";
    input: typeof WorktreeServiceListRequestSchema;
    output: typeof WorktreeServiceListResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Create
   */
  create: {
    methodKind: "unary";
    input: typeof WorktreeServiceCreateRequestSchema;
    output: typeof WorktreeServiceCreateResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.ListArchives
   */
  listArchives: {
    methodKind: "unary";
    input: typeof WorktreeServiceListArchivesRequestSchema;
    output: typeof WorktreeServiceListArchivesResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Restore
   */
  restore: {
    methodKind: "unary";
    input: typeof WorktreeServiceRestoreRequestSchema;
    output: typeof WorktreeServiceRestoreResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Ps
   */
  ps: {
    methodKind: "unary";
    input: typeof WorktreeServicePsRequestSchema;
    output: typeof WorktreeServicePsResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Show
   */
  show: {
    methodKind: "unary";
    input: typeof WorktreeServiceShowRequestSchema;
    output: typeof WorktreeServiceShowResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Sleep
   */
  sleep: {
    methodKind: "unary";
    input: typeof WorktreeServiceSleepRequestSchema;
    output: typeof WorktreeServiceSleepResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Activate
   */
  activate: {
    methodKind: "unary";
    input: typeof WorktreeServiceActivateRequestSchema;
    output: typeof WorktreeServiceActivateResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Remove
   */
  remove: {
    methodKind: "unary";
    input: typeof WorktreeServiceRemoveRequestSchema;
    output: typeof WorktreeServiceRemoveResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.ForceDeleteBranch
   */
  forceDeleteBranch: {
    methodKind: "unary";
    input: typeof WorktreeServiceForceDeleteBranchRequestSchema;
    output: typeof WorktreeServiceForceDeleteBranchResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.Set
   */
  set: {
    methodKind: "unary";
    input: typeof WorktreeServiceSetRequestSchema;
    output: typeof WorktreeServiceSetResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.PersistSortOrder
   */
  persistSortOrder: {
    methodKind: "unary";
    input: typeof WorktreeServicePersistSortOrderRequestSchema;
    output: typeof WorktreeServicePersistSortOrderResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.DetectedList
   */
  detectedList: {
    methodKind: "unary";
    input: typeof WorktreeServiceDetectedListRequestSchema;
    output: typeof WorktreeServiceDetectedListResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.LineageList
   */
  lineageList: {
    methodKind: "unary";
    input: typeof WorktreeServiceLineageListRequestSchema;
    output: typeof WorktreeServiceLineageListResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.PrefetchCreateBase
   */
  prefetchCreateBase: {
    methodKind: "unary";
    input: typeof WorktreeServicePrefetchCreateBaseRequestSchema;
    output: typeof WorktreeServicePrefetchCreateBaseResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.ResolvePrBase
   */
  resolvePrBase: {
    methodKind: "unary";
    input: typeof WorktreeServiceResolvePrBaseRequestSchema;
    output: typeof WorktreeServiceResolvePrBaseResponseSchema;
  },
  /**
   * @generated from rpc agent_start.runtime.v1.WorktreeService.BranchRenameFailureOutput
   */
  branchRenameFailureOutput: {
    methodKind: "unary";
    input: typeof WorktreeServiceBranchRenameFailureOutputRequestSchema;
    output: typeof WorktreeServiceBranchRenameFailureOutputResponseSchema;
  },
  /**
   * Why: base-drift banners must reflect every repo
   * on the host, so the subscription is host-scoped rather than per-worktree.
   *
   * @generated from rpc agent_start.runtime.v1.WorktreeService.SubscribeStateEvents
   */
  subscribeStateEvents: {
    methodKind: "server_streaming";
    input: typeof WorktreeServiceSubscribeStateEventsRequestSchema;
    output: typeof WorktreeServiceSubscribeStateEventsResponseSchema;
  },
}>;